requirements.txt
aenum==3.1.15
ai-api-client-sdk==1.28.0
ai-core-sdk==1.22.3
arabic-reshaper==2.1.3
asgiref==3.6.0
asn1crypto==1.5.1
astroid==3.0.1
async-timeout==4.0.2
Babel==2.9.1
backports.zoneinfo==0.2.1
black==22.12.0
boto3==1.21.46
botocore==1.24.46
build==1.0.3
bureau==0.0.1
cachetools==5.3.2
certifi==2022.5.18.1
cffi==1.15.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.0.12
click==8.1.7
cmake==3.27.9
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==40.0.1
cssselect2==0.6.0
Cython==0.29.28
datetime-truncate==1.1.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20200713
dill==0.3.7
Django==4.2.10
django-active-link==0.1.8
django-admin-charts==1.3.0
django-admin-interface==0.19.0
django-admin-relation-links==0.2.5
django-admin-tools==0.9.2
django-ajax-datatable==4.4.5
django-appconf==1.0.6
django-baton==2.3.0
django-bower==5.2.0
django-browser-reload==1.7.0
django-channels==0.7.0
django-ckeditor==6.7.0
django-clearcache==1.2.1
django-colorfield==0.7.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-crum==0.7.9
django-csp-plus==3.0.1
django-daterange-filterspec==2.0.5
django-debug-toolbar==4.2.0
django-dump-die==0.1.7
django-enumfield==3.1
django-environ==0.8.1
django-extensions==3.1.3
django-filter==23.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==3.0.3
django-imagekit==5.0.0
django-import-export==3.1.0
django-jazzmin==2.6.0
django-jet-reboot==1.3.5
django-js-asset==2.2.0
django-json-widget==1.1.1
django-memoize==2.3.1
django-model-utils==4.3.1
django-multiselectfield==0.1.12
django-nested-inlines==0.1
django-nonrelated-inlines==0.2
django-nvd3==0.9.7
django-pandas==0.6.6
django-qr-code==3.1.1
django-qrcode==0.3
django-relativedelta==2.0.0
django-rest-swagger==2.2.0
django-session-timeout==0.1.0
django-taggit==5.0.1
django-templated-mail==1.1.1
django-tinymce==3.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.1.0
fpdf==1.7.2
future==0.18.2
google-auth==2.25.2
gopay==2.0.1
html5lib==1.1
idna==3.3
image==1.5.33
import-export==0.2.67.dev6
importlib-metadata==5.0.0
ipapi==1.0.4
isort==5.13.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
Js2Py==0.74
kubernetes==28.1.0
lxml==4.8.0
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.1
marshmallow==3.15.0
mccabe==0.7.0
metaflow==2.8.0
msgpack==1.0.5
mypy-extensions==1.0.0
mysqlclient @ file:///C:/Users/hp/desktop/mysqlclient-1.4.6-cp39-cp39-win32.whl#sha256=7e0fb8aa09b57b3d92d8484f3ad507d1c7cc74de5b41946ea704d4d1907b0fe4
ndg-httpsclient==0.5.1
networkx==2.6.3
nltk==3.7
num2words==0.5.12
numpy==1.22.3
oauthlib==3.2.2
odfpy==1.4.1
onesignal-python-api==2.0.0
openapi-codec==1.3.2
openpyxl==3.0.9
oscrypto==1.3.0
packaging==21.3
panda==0.3.1
pandas==1.4.3
pathspec==0.12.1
pep517==0.13.0
pilkit==3.0
Pillow==9.3.0
pip-tools==6.10.0
platformdirs==4.1.0
polib==1.1.1
psutil==5.9.5
psycopg2-binary==2.9.5
pusher==3.3.2
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycel==1.0b30
pycparser==2.21
pydantic==1.10.14
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyhumps==3.8.0
pyjsparser==2.7.1
PyJWT==2.5.0
pylint==3.0.3
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyotp==2.9.0
pyparsing==3.0.9
PyPDF2==3.0.1
PyPDF3==1.0.6
pypng==0.20220715.0
pyproject_hooks==1.0.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.2
python-environ==0.4.54
python-gettext==5.0
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==5.4.1
qrcode==7.3.1
redis==4.5.4
regex==2022.6.2
register==0.1
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
sap-ai-core-metaflow==1.1.14
segno==1.5.2
self==2020.12.3
simplejson==3.18.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==5.5.0
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.64.0
typing_extensions==4.4.0
tzdata==2022.1
tzlocal==4.2
Unidecode==1.3.6
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
websocket-client==1.7.0
whitenoise==6.2.0
wrapt==1.16.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwings==0.28.5
xlwt==1.3.0
zipp==3.9.0